🔧 Modify : non-L2 functions to L0
